+++ lib/models/money_objects/rental_unit/rental_units.dart
@@ -12,8 +12,8 @@ class RentUnits extends MoneyObjects<RentUnit> {
   }
 
   @override
-  void loadFromJson(final List<Json> rows) {
-    for (final Json row in rows) {
+  void loadFromJson(final List<MyJson> rows) {
+    for (final MyJson row in rows) {
       addEntry(RentUnit.fromSqlite(row));
     }
   }